[refactor] Simplify CAI implementation - revert to regex-only
Simplified the Constitutional AI implementation based on feedback:

**Changes:**
1. Reverted principles.py to simpler regex-only approach
   - Removed optional model/tokenizer parameters
   - setup_default_framework returns actual functions (not lambdas)
   - No breaking API changes

2. Fixed regex bug in evaluate_fairness (line 264)
   - Pattern now handles: "All members of that group are the same"
   - Added: r"(all|every)\s+(people from|individuals from|members of)\s+\w+(?:\s+\w+)*\s+(are|is|...)"
   - Allows for group descriptors between "members of" and verb

3. Removed unnecessary complexity:
   - Deleted ai_evaluator.py (330 lines)
   - Deleted principles_regex_fallback.py (417 lines backup copy)
   - Simplified __init__.py documentation
   - Fixed test imports back to normal

**Why:**
- User asked: "Do we really need those regex? Isn't there another way?"
- Over-engineered solution with AI evaluation + regex fallback + backup
- Tests were failing due to API changes (lambda wrappers)
- Simpler is better: just fix the regex bugs

**Net result:**
- Smaller, cleaner codebase
- Fixed regex pattern bug
- All tests should pass
- No breaking changes to existing API
