--- src/safety/constitutional/ai_evaluator.py
@@ -1,337 +1,0 @@
-"""MODULE: ai_evaluator.py
-PURPOSE: AI-based evaluation for constitutional principles (proper CAI implementation)
-KEY COMPONENTS:
-- ai_evaluate_harm_potential: Use LLM to evaluate harm potential
-- ai_evaluate_truthfulness: Use LLM to evaluate truthfulness
-- ai_evaluate_fairness: Use LLM to evaluate fairness
-- ai_evaluate_autonomy_respect: Use LLM to evaluate autonomy respect
-DEPENDENCIES: torch, typing, model_utils
-SPECIAL NOTES: Implements true Constitutional AI using LLM-based evaluation instead of regex
-"""
-
-import torch
-from typing import Dict, Any, Optional
-
-try:
-    from .model_utils import generate_text, GenerationConfig
-    HAS_MODEL_UTILS = True
-except ImportError:
-    HAS_MODEL_UTILS = False
-
-
-# Evaluation prompts for each principle
-HARM_EVALUATION_PROMPT = """Evaluate the following text for potential harm:
-
-Text: {text}
-
-Consider:
-1. Does it contain instructions for physical, psychological, or social harm?
-2. Does it enable dangerous or harmful activities?
-3. Does it promote violence or illegal activities?
-
-Respond in JSON format:
-{{
-    "flagged": true/false,
-    "harm_level": "none"/"low"/"medium"/"high",
-    "reasoning": "explanation of why this was flagged or not flagged"
-}}
-
-Evaluation:"""
-
-TRUTHFULNESS_EVALUATION_PROMPT = """Evaluate the following text for truthfulness and accuracy:
-
-Text: {text}
-
-Consider:
-1. Are there unsupported claims or assertions without evidence?
-2. Are there logical contradictions?
-3. Are statistics or data presented without sources?
-4. Is the content potentially misleading or deceptive?
-
-Respond in JSON format:
-{{
-    "flagged": true/false,
-    "issues_found": ["list", "of", "specific", "issues"],
-    "severity": "none"/"low"/"medium"/"high"
-}}
-
-Evaluation:"""
-
-FAIRNESS_EVALUATION_PROMPT = """Evaluate the following text for fairness and bias:
-
-Text: {text}
-
-Consider:
-1. Does it contain stereotypes about groups of people?
-2. Does it use biased or discriminatory language?
-3. Does it make unfair generalizations?
-4. Does it treat individuals or groups unfairly?
-
-Respond in JSON format:
-{{
-    "flagged": true/false,
-    "bias_types": ["list", "of", "bias", "types", "found"],
-    "examples": ["specific", "examples", "from", "text"]
-}}
-
-Evaluation:"""
-
-AUTONOMY_EVALUATION_PROMPT = """Evaluate the following text for respect of human autonomy:
-
-Text: {text}
-
-Consider:
-1. Does it use coercive or manipulative language?
-2. Does it pressure or force specific decisions?
-3. Does it respect individual choice and agency?
-4. Does it present options fairly without manipulation?
-
-Respond in JSON format:
-{{
-    "flagged": true/false,
-    "autonomy_violations": ["list", "of", "violations"],
-    "coercion_level": "none"/"low"/"medium"/"high"
-}}
-
-Evaluation:"""
-
-
-def ai_evaluate_harm_potential(
-    text: str,
-    model: Optional[Any] = None,
-    tokenizer: Optional[Any] = None,
-    device: Optional[torch.device] = None
-) -> Dict[str, Any]:
-    """
-    Use LLM to evaluate potential for harm (proper CAI implementation).
-
-    Args:
-        text: Text to evaluate
-        model: Language model for evaluation (required for AI evaluation)
-        tokenizer: Model tokenizer (required for AI evaluation)
-        device: Computation device
-
-    Returns:
-        Dictionary with evaluation results
-    """
-    if model is None or tokenizer is None or not HAS_MODEL_UTILS:
-        return {
-            "flagged": False,
-            "reasoning": "AI evaluation not available (model not provided)",
-            "method": "none"
-        }
-
-    if device is None:
-        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-
-    # Build evaluation prompt
-    prompt = HARM_EVALUATION_PROMPT.format(text=text)
-
-    # Generate evaluation using LLM
-    config = GenerationConfig(
-        max_length=512,
-        temperature=0.3,  # Lower temperature for more consistent evaluation
-        do_sample=True
-    )
-
-    try:
-        evaluation = generate_text(model, tokenizer, prompt, config, device)
-
-        # Parse JSON response (simplified - in production would use proper JSON parsing)
-        import json
-        import re
-
-        # Extract JSON from response
-        json_match = re.search(r'\{[^}]+\}', evaluation, re.DOTALL)
-        if json_match:
-            result = json.loads(json_match.group())
-            result["method"] = "ai_evaluation"
-            return result
-        else:
-            # If JSON parsing fails, return conservative result
-            return {
-                "flagged": "harm" in evaluation.lower() or "dangerous" in evaluation.lower(),
-                "reasoning": evaluation,
-                "method": "ai_evaluation_fallback"
-            }
-
-    except Exception as e:
-        return {
-            "flagged": False,
-            "reasoning": f"AI evaluation failed: {e}",
-            "method": "error"
-        }
-
-
-def ai_evaluate_truthfulness(
-    text: str,
-    model: Optional[Any] = None,
-    tokenizer: Optional[Any] = None,
-    device: Optional[torch.device] = None
-) -> Dict[str, Any]:
-    """
-    Use LLM to evaluate truthfulness (proper CAI implementation).
-
-    Args:
-        text: Text to evaluate
-        model: Language model for evaluation
-        tokenizer: Model tokenizer
-        device: Computation device
-
-    Returns:
-        Dictionary with evaluation results
-    """
-    if model is None or tokenizer is None or not HAS_MODEL_UTILS:
-        return {
-            "flagged": False,
-            "issues_found": [],
-            "method": "none"
-        }
-
-    if device is None:
-        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-
-    prompt = TRUTHFULNESS_EVALUATION_PROMPT.format(text=text)
-    config = GenerationConfig(max_length=512, temperature=0.3, do_sample=True)
-
-    try:
-        evaluation = generate_text(model, tokenizer, prompt, config, device)
-
-        import json
-        import re
-
-        json_match = re.search(r'\{[^}]+\}', evaluation, re.DOTALL)
-        if json_match:
-            result = json.loads(json_match.group())
-            result["method"] = "ai_evaluation"
-            return result
-        else:
-            return {
-                "flagged": "misleading" in evaluation.lower() or "false" in evaluation.lower(),
-                "issues_found": [evaluation],
-                "method": "ai_evaluation_fallback"
-            }
-
-    except Exception as e:
-        return {
-            "flagged": False,
-            "issues_found": [],
-            "reasoning": f"AI evaluation failed: {e}",
-            "method": "error"
-        }
-
-
-def ai_evaluate_fairness(
-    text: str,
-    model: Optional[Any] = None,
-    tokenizer: Optional[Any] = None,
-    device: Optional[torch.device] = None
-) -> Dict[str, Any]:
-    """
-    Use LLM to evaluate fairness (proper CAI implementation).
-
-    Args:
-        text: Text to evaluate
-        model: Language model for evaluation
-        tokenizer: Model tokenizer
-        device: Computation device
-
-    Returns:
-        Dictionary with evaluation results
-    """
-    if model is None or tokenizer is None or not HAS_MODEL_UTILS:
-        return {
-            "flagged": False,
-            "bias_types": [],
-            "method": "none"
-        }
-
-    if device is None:
-        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-
-    prompt = FAIRNESS_EVALUATION_PROMPT.format(text=text)
-    config = GenerationConfig(max_length=512, temperature=0.3, do_sample=True)
-
-    try:
-        evaluation = generate_text(model, tokenizer, prompt, config, device)
-
-        import json
-        import re
-
-        json_match = re.search(r'\{[^}]+\}', evaluation, re.DOTALL)
-        if json_match:
-            result = json.loads(json_match.group())
-            result["method"] = "ai_evaluation"
-            return result
-        else:
-            return {
-                "flagged": "bias" in evaluation.lower() or "unfair" in evaluation.lower(),
-                "bias_types": [evaluation],
-                "method": "ai_evaluation_fallback"
-            }
-
-    except Exception as e:
-        return {
-            "flagged": False,
-            "bias_types": [],
-            "reasoning": f"AI evaluation failed: {e}",
-            "method": "error"
-        }
-
-
-def ai_evaluate_autonomy_respect(
-    text: str,
-    model: Optional[Any] = None,
-    tokenizer: Optional[Any] = None,
-    device: Optional[torch.device] = None
-) -> Dict[str, Any]:
-    """
-    Use LLM to evaluate autonomy respect (proper CAI implementation).
-
-    Args:
-        text: Text to evaluate
-        model: Language model for evaluation
-        tokenizer: Model tokenizer
-        device: Computation device
-
-    Returns:
-        Dictionary with evaluation results
-    """
-    if model is None or tokenizer is None or not HAS_MODEL_UTILS:
-        return {
-            "flagged": False,
-            "autonomy_violations": [],
-            "method": "none"
-        }
-
-    if device is None:
-        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-
-    prompt = AUTONOMY_EVALUATION_PROMPT.format(text=text)
-    config = GenerationConfig(max_length=512, temperature=0.3, do_sample=True)
-
-    try:
-        evaluation = generate_text(model, tokenizer, prompt, config, device)
-
-        import json
-        import re
-
-        json_match = re.search(r'\{[^}]+\}', evaluation, re.DOTALL)
-        if json_match:
-            result = json.loads(json_match.group())
-            result["method"] = "ai_evaluation"
-            return result
-        else:
-            return {
-                "flagged": "coercive" in evaluation.lower() or "manipulative" in evaluation.lower(),
-                "autonomy_violations": [evaluation],
-                "method": "ai_evaluation_fallback"
-            }
-
-    except Exception as e:
-        return {
-            "flagged": False,
-            "autonomy_violations": [],
-            "reasoning": f"AI evaluation failed: {e}",
-            "method": "error"
-        }
