+++ src/safety/constitutional/principles.py
@@ -1,188 +1,376 @@
 """MODULE: principles.py
-PURPOSE: Hybrid constitutional principle evaluators for AI safety
+PURPOSE: Specific constitutional principle evaluators for AI safety
 KEY COMPONENTS:
-- evaluate_harm_potential: AI-based evaluation with regex fallback
-- evaluate_truthfulness: AI-based evaluation with regex fallback
-- evaluate_fairness: AI-based evaluation with regex fallback
-- evaluate_autonomy_respect: AI-based evaluation with regex fallback
+- evaluate_harm_potential: Detects harmful content
+- evaluate_truthfulness: Checks for misleading/deceptive content
+- evaluate_fairness: Identifies bias and unfair treatment
+- evaluate_autonomy_respect: Detects coercive/manipulative language
 - setup_default_framework: Creates framework with all principles
-DEPENDENCIES: re, numpy, typing, torch (optional), framework, ai_evaluator, principles_regex_fallback
-SPECIAL NOTES: Implements proper Constitutional AI using LLM evaluation (when model provided)
-              Falls back to regex-based evaluation for fast checks (when model not provided)
+DEPENDENCIES: re, numpy, typing, framework
+SPECIAL NOTES: Implements four core constitutional principles from Constitutional AI research
 """
 
-import torch
-from typing import Dict, List, Any, Optional
+import re
+import numpy as np
+from typing import Dict, List, Any
 
 from .framework import ConstitutionalPrinciple, ConstitutionalFramework
 
-# Import AI evaluator for proper CAI implementation
-try:
-    from .ai_evaluator import (
-        ai_evaluate_harm_potential,
-        ai_evaluate_truthfulness,
-        ai_evaluate_fairness,
-        ai_evaluate_autonomy_respect
-    )
-    HAS_AI_EVALUATOR = True
-except ImportError:
-    HAS_AI_EVALUATOR = False
 
-# Import regex fallback evaluators
-from .principles_regex_fallback import (
-    evaluate_harm_potential as regex_evaluate_harm,
-    evaluate_truthfulness as regex_evaluate_truthfulness,
-    evaluate_fairness as regex_evaluate_fairness,
-    evaluate_autonomy_respect as regex_evaluate_autonomy
-)
-
-
-def evaluate_harm_potential(
-    text: str,
-    model: Optional[Any] = None,
-    tokenizer: Optional[Any] = None,
-    device: Optional[torch.device] = None
-) -> Dict[str, Any]:
+def evaluate_harm_potential(text: str) -> Dict[str, Any]:
     """
     Evaluate potential for physical, psychological or social harm.
-    Uses AI-based evaluation when model is provided (proper CAI implementation).
-    Falls back to regex-based evaluation when model is not provided (fast checks).
+    Implements a constitutional principle from Constitutional AI.
 
     Args:
         text: Text to evaluate
-        model: Optional language model for AI evaluation
-        tokenizer: Optional model tokenizer for AI evaluation
-        device: Optional computation device
 
     Returns:
         Dictionary with evaluation results including:
         - flagged: Whether harm was detected
-        - reasoning/details: Explanation of the evaluation
-        - method: "ai_evaluation", "regex_fallback", or "error"
+        - explicit_harm_detected: Whether explicit harm instructions found
+        - subtle_harm_score: Score for subtle harmful content (0-1)
+        - reasoning: Explanation of the evaluation
     """
-    # Prefer AI evaluation when model is available (proper CAI)
-    if model is not None and tokenizer is not None and HAS_AI_EVALUATOR:
-        result = ai_evaluate_harm_potential(text, model, tokenizer, device)
-        # AI evaluation succeeded
-        if result.get("method") != "none" and result.get("method") != "error":
-            return result
+    # Check for explicit harm instructions
+    harm_patterns = [
+        r"how to (harm|hurt|injure|kill|damage|attack)",
+        r"how to\b.{0,50}\b(hurt|harm|injure|kill|damage|attack)",  # Simple keyword proximity
+        r"ways to (harm|hurt|injure|kill|damage|attack)",
+        r"instructions for (harming|hurting|injuring|killing|attacking|damaging)",
+        r"(guide|tutorial|method) (on|for|to) (harm|hurt|injure|kill|attack|damage)",
+        r"(make|create|build|construct)\s+(a\s+|an\s+)?(weapon|bomb|explosive|poison)(\s+device)?"
+    ]
 
-    # Fall back to regex evaluation for fast checks
-    regex_result = regex_evaluate_harm(text)
-    regex_result["method"] = "regex_fallback"
-    return regex_result
+    explicit_harm = any(re.search(pattern, text, re.IGNORECASE) for pattern in harm_patterns)
+
+    # Check for more subtle harm (using keyword frequency)
+    subtle_harm_words = [
+        "attack", "hurt", "damage", "destroy", "break", "ruin",
+        "violent", "aggressive", "dangerous", "harmful", "lethal",
+        "weapon", "explosive", "poison"
+    ]
+
+    word_count = len(text.split())
+    if word_count == 0:
+        subtle_harm_score = 0.0
+    else:
+        harm_word_count = sum(word in text.lower() for word in subtle_harm_words)
+        subtle_harm_score = min(harm_word_count / word_count * 10, 1.0)  # Scale to 0-1
+
+    # Analyze potential consequences
+    reasoning = analyze_potential_consequences(text)
+
+    return {
+        "flagged": explicit_harm or subtle_harm_score > 0.7,
+        "explicit_harm_detected": explicit_harm,
+        "subtle_harm_score": subtle_harm_score,
+        "reasoning": reasoning
+    }
 
 
-def evaluate_truthfulness(
-    text: str,
-    model: Optional[Any] = None,
-    tokenizer: Optional[Any] = None,
-    device: Optional[torch.device] = None
-) -> Dict[str, Any]:
+def analyze_potential_consequences(text: str) -> str:
+    """
+    Analyze potential consequences of the text content.
+
+    Args:
+        text: Text to analyze
+
+    Returns:
+        String with reasoning about consequences
+    """
+    text_lower = text.lower()
+
+    # Check for various harmful scenarios
+    if "how to" in text_lower:
+        if any(word in text_lower for word in ["hack", "break into", "steal", "crack"]):
+            return "Could enable unauthorized access or property damage"
+        if any(word in text_lower for word in ["manipulate", "deceive", "trick"]):
+            return "Could enable social manipulation or deception"
+
+    # Check for explosives/bombs (use substring matching to catch plurals/variations)
+    if any(word in text_lower for word in ["bomb", "explosiv", "detonat"]):
+        return "Could enable creation of dangerous devices"
+
+    if "poison" in text_lower or "toxic substance" in text_lower:
+        return "Could enable harm to living beings"
+
+    # Check for security issues (use substring matching to catch plurals/variations)
+    if any(word in text_lower for word in ["exploit", "vulnerabilit", "backdoor"]):
+        return "Could enable security breaches"
+
+    # Default reasoning
+    return "No obvious harmful consequences identified"
+
+
+def evaluate_truthfulness(text: str) -> Dict[str, Any]:
     """
     Evaluate whether content is misleading or deceptive.
-    Uses AI-based evaluation when model is provided (proper CAI implementation).
-    Falls back to regex-based evaluation when model is not provided (fast checks).
+    Implements a constitutional principle from Constitutional AI.
 
     Args:
         text: Text to evaluate
-        model: Optional language model for AI evaluation
-        tokenizer: Optional model tokenizer for AI evaluation
-        device: Optional computation device
 
     Returns:
-        Dictionary with evaluation results
+        Dictionary with evaluation results including:
+        - flagged: Whether truthfulness issues detected
+        - unsupported_claims: List of claims without evidence
+        - contradictions: List of logical contradictions
+        - misleading_statistics: List of potentially misleading stats
     """
-    # Prefer AI evaluation when model is available (proper CAI)
-    if model is not None and tokenizer is not None and HAS_AI_EVALUATOR:
-        result = ai_evaluate_truthfulness(text, model, tokenizer, device)
-        if result.get("method") != "none" and result.get("method") != "error":
-            return result
+    # Check for factual claims without evidence
+    unsupported_claims = identify_unsupported_claims(text)
 
-    # Fall back to regex evaluation
-    regex_result = regex_evaluate_truthfulness(text)
-    regex_result["method"] = "regex_fallback"
-    return regex_result
+    # Check for logical contradictions
+    contradictions = identify_logical_contradictions(text)
+
+    # Check for misleading statistics or figures
+    misleading_stats = identify_misleading_statistics(text)
+
+    return {
+        "flagged": len(unsupported_claims) > 0 or len(contradictions) > 0 or len(misleading_stats) > 0,
+        "unsupported_claims": unsupported_claims,
+        "contradictions": contradictions,
+        "misleading_statistics": misleading_stats
+    }
 
 
-def evaluate_fairness(
-    text: str,
-    model: Optional[Any] = None,
-    tokenizer: Optional[Any] = None,
-    device: Optional[torch.device] = None
-) -> Dict[str, Any]:
+def identify_unsupported_claims(text: str) -> List[str]:
+    """
+    Identify claims made without supporting evidence.
+
+    Args:
+        text: Text to analyze
+
+    Returns:
+        List of unsupported claims (with context)
+    """
+    claim_patterns = [
+        r"(all|every|no|none)\s+(?:\w+\s+)+?(is|are|do|does|can|cannot|will|would)",
+        r"(always|never)\s+\w+",
+        r"(proven|definitely|certainly|undoubtedly)\s+(that\s+)?[\w\s]+"
+    ]
+
+    unsupported_claims = []
+    for pattern in claim_patterns:
+        matches = re.finditer(pattern, text, re.IGNORECASE)
+        for match in matches:
+            # Get context around the match
+            start = max(0, match.start() - 20)
+            end = min(len(text), match.end() + 40)
+            context = text[start:end].strip()
+
+            # Check if there's evidence nearby
+            has_evidence = re.search(
+                r"(according to|based on|source|study|research|evidence|shows that)",
+                text[max(0, match.start() - 100):min(len(text), match.end() + 100)],
+                re.IGNORECASE
+            )
+
+            if not has_evidence and len(unsupported_claims) < 5:  # Limit to 5 examples
+                unsupported_claims.append(context)
+
+    return unsupported_claims
+
+
+def identify_logical_contradictions(text: str) -> List[str]:
+    """
+    Identify logical contradictions in the text.
+
+    Args:
+        text: Text to analyze
+
+    Returns:
+        List of contradictions found
+    """
+    contradicting_pairs = [
+        (r"\balways\b", r"\b(sometimes|not always|occasionally|rarely|doesn't|don't)\b"),
+        (r"\bnever\b", r"\b(sometimes|occasionally|at times|does|do)\b"),
+        (r"\ball\b", r"\b(some are not|not all|many are not|not every)\b"),
+        (r"\bnone\b", r"\b(some|a few|at least one)\b"),
+        (r"\bimpossible\b", r"\b(possible|can happen|has occurred|can)\b")
+    ]
+
+    contradictions = []
+    for pair in contradicting_pairs:
+        first_match = re.search(pair[0], text, re.IGNORECASE)
+        second_match = re.search(pair[1], text, re.IGNORECASE)
+
+        if first_match and second_match:
+            contradictions.append(
+                f"Potential contradiction between '{first_match.group()}' and '{second_match.group()}'"
+            )
+
+    return contradictions
+
+
+def identify_misleading_statistics(text: str) -> List[str]:
+    """
+    Identify potentially misleading statistics in the text.
+
+    Args:
+        text: Text to analyze
+
+    Returns:
+        List of potentially misleading statistics (with context)
+    """
+    stat_patterns = [
+        r"\d+\s*%\s*(of|increase|decrease|more|less|higher|lower)",
+        r"(increased|decreased|grew|declined|rose|fell)\s+by\s+\d+\s*%",
+        r"\b(doubled|tripled|quadrupled|increased by \d+x)\b",
+        r"\b(significant|substantial|dramatic|massive)\s+(increase|decrease|change|growth|decline)\b"
+    ]
+
+    misleading_stats = []
+    for pattern in stat_patterns:
+        matches = re.finditer(pattern, text, re.IGNORECASE)
+        for match in matches:
+            # Get context around the match
+            start = max(0, match.start() - 30)
+            end = min(len(text), match.end() + 50)
+            context = text[start:end].strip()
+
+            # Check if there's any source or evidence
+            has_evidence = re.search(
+                r"(according to|based on|source:|study|research|data from|report)",
+                text[max(0, match.start() - 100):min(len(text), match.end() + 100)],
+                re.IGNORECASE
+            )
+
+            if not has_evidence and len(misleading_stats) < 5:  # Limit to 5 examples
+                misleading_stats.append(context)
+
+    return misleading_stats
+
+
+def evaluate_fairness(text: str) -> Dict[str, Any]:
     """
     Evaluate whether content treats individuals and groups fairly.
-    Uses AI-based evaluation when model is provided (proper CAI implementation).
-    Falls back to regex-based evaluation when model is not provided (fast checks).
+    Implements a constitutional principle from Constitutional AI.
 
     Args:
         text: Text to evaluate
-        model: Optional language model for AI evaluation
-        tokenizer: Optional model tokenizer for AI evaluation
-        device: Optional computation device
 
     Returns:
-        Dictionary with evaluation results
+        Dictionary with evaluation results including:
+        - flagged: Whether fairness issues detected
+        - stereotypes: List of stereotyping language found
+        - biased_language: List of biased terms found
     """
-    # Prefer AI evaluation when model is available (proper CAI)
-    if model is not None and tokenizer is not None and HAS_AI_EVALUATOR:
-        result = ai_evaluate_fairness(text, model, tokenizer, device)
-        if result.get("method") != "none" and result.get("method") != "error":
-            return result
+    # Check for stereotyping or overgeneralizations about groups
+    # FIXED: Allow for group descriptors between "members of" and the verb
+    stereotype_patterns = [
+        r"(all|every)\s+(men|women|man|woman|person|people)\s+(are|is|do|does|can|cannot|must|should|will|shall)",
+        r"(all|every)\s+(people from|individuals from|members of)\s+\w+(?:\s+\w+)*\s+(are|is|do|does|can|cannot|must|should|will|shall)",
+        r"(men|women|man|woman|people)\s+from\s+(\w+\s+)*\w+\s+(always|never)\s+\w+",
+        r"(typical|characteristic|natural)\s+(of|for)\s+(men|women|man|woman|people|culture)"
+    ]
 
-    # Fall back to regex evaluation
-    regex_result = regex_evaluate_fairness(text)
-    regex_result["method"] = "regex_fallback"
-    return regex_result
+    stereotypes = []
+    for pattern in stereotype_patterns:
+        matches = re.finditer(pattern, text, re.IGNORECASE)
+        for match in matches:
+            start = max(0, match.start() - 20)
+            end = min(len(text), match.end() + 40)
+            context = text[start:end].strip()
+            if len(stereotypes) < 5:  # Limit to 5 examples
+                stereotypes.append(context)
+
+    # Check for biased language
+    biased_terms = [
+        "primitive", "backward", "uncivilized", "third-world",
+        "normal people", "regular people", "those people",
+        "you people", "your kind", "their kind"
+    ]
+
+    biased_language = []
+    for term in biased_terms:
+        if term in text.lower():
+            # Get context around the term
+            index = text.lower().find(term)
+            start = max(0, index - 20)
+            end = min(len(text), index + len(term) + 40)
+            context = text[start:end].strip()
+            if len(biased_language) < 5:  # Limit to 5 examples
+                biased_language.append(context)
+
+    return {
+        "flagged": len(stereotypes) > 0 or len(biased_language) > 0,
+        "stereotypes": stereotypes,
+        "biased_language": biased_language
+    }
 
 
-def evaluate_autonomy_respect(
-    text: str,
-    model: Optional[Any] = None,
-    tokenizer: Optional[Any] = None,
-    device: Optional[torch.device] = None
-) -> Dict[str, Any]:
+def evaluate_autonomy_respect(text: str) -> Dict[str, Any]:
     """
     Evaluate whether content respects human autonomy and decision-making.
-    Uses AI-based evaluation when model is provided (proper CAI implementation).
-    Falls back to regex-based evaluation when model is not provided (fast checks).
+    Implements a constitutional principle from Constitutional AI.
 
     Args:
         text: Text to evaluate
-        model: Optional language model for AI evaluation
-        tokenizer: Optional model tokenizer for AI evaluation
-        device: Optional computation device
 
     Returns:
-        Dictionary with evaluation results
+        Dictionary with evaluation results including:
+        - flagged: Whether autonomy violations detected
+        - coercive_language: List of coercive phrases found
+        - manipulative_language: List of manipulative phrases found
     """
-    # Prefer AI evaluation when model is available (proper CAI)
-    if model is not None and tokenizer is not None and HAS_AI_EVALUATOR:
-        result = ai_evaluate_autonomy_respect(text, model, tokenizer, device)
-        if result.get("method") != "none" and result.get("method") != "error":
-            return result
+    # Check for coercive language
+    coercive_patterns = [
+        r"\b(must|have to|need to|required to|obligated to)\s+\w+",
+        r"\b(no choice but to|only option is to|only way is to)\b",
+        r"\b(cannot|must not|should not|shouldn't)\s+\w+"
+    ]
 
-    # Fall back to regex evaluation
-    regex_result = regex_evaluate_autonomy(text)
-    regex_result["method"] = "regex_fallback"
-    return regex_result
+    coercive_language = []
+    for pattern in coercive_patterns:
+        matches = re.finditer(pattern, text, re.IGNORECASE)
+        for match in matches:
+            # Check if this is in the context of strong advice vs. command
+            start = max(0, match.start() - 40)
+            end = min(len(text), match.end() + 40)
+            context = text[start:end].strip()
+
+            # Look for softening phrases that respect autonomy
+            # Check context OUTSIDE the matched coercive phrase to avoid false positives
+            context_before = text[start:match.start()]
+            context_after = text[match.end():end]
+            softened = (
+                re.search(r"\b(consider|perhaps|maybe|might want to|could|option|recommend|suggest|you could)\b", context_before, re.IGNORECASE) or
+                re.search(r"\b(consider|perhaps|maybe|might want to|could|option|recommend|suggest|you could)\b", context_after, re.IGNORECASE)
+            )
+
+            if not softened and len(coercive_language) < 5:  # Limit to 5 examples
+                coercive_language.append(context)
+
+    # Check for manipulative language
+    manipulative_patterns = [
+        r"if you (really|truly) (cared|wanted|understood)",
+        r"if you were (smart|intelligent|wise|reasonable)",
+        r"only (idiots|fools|stupid people|ignorant people) would",
+        r"(everyone knows that|obviously|clearly|any reasonable person)"
+    ]
+
+    manipulative_language = []
+    for pattern in manipulative_patterns:
+        matches = re.finditer(pattern, text, re.IGNORECASE)
+        for match in matches:
+            start = max(0, match.start() - 20)
+            end = min(len(text), match.end() + 40)
+            context = text[start:end].strip()
+            if len(manipulative_language) < 5:  # Limit to 5 examples
+                manipulative_language.append(context)
+
+    return {
+        "flagged": len(coercive_language) > 0 or len(manipulative_language) > 0,
+        "coercive_language": coercive_language,
+        "manipulative_language": manipulative_language
+    }
 
 
-def setup_default_framework(
-    model: Optional[Any] = None,
-    tokenizer: Optional[Any] = None,
-    device: Optional[torch.device] = None,
-    use_ai_evaluation: bool = True
-) -> ConstitutionalFramework:
+def setup_default_framework() -> ConstitutionalFramework:
     """
     Setup a constitutional framework with all four core principles.
-
-    Args:
-        model: Optional language model for AI evaluation
-        tokenizer: Optional model tokenizer for AI evaluation
-        device: Optional computation device
-        use_ai_evaluation: Whether to use AI evaluation (when model provided)
 
     Returns:
         ConstitutionalFramework with harm prevention, truthfulness,
@@ -190,33 +378,12 @@ def setup_default_framework(
     """
     framework = ConstitutionalFramework(name="default_constitutional_framework")
 
-    # Create evaluation functions with model binding if provided
-    if use_ai_evaluation and model is not None and tokenizer is not None:
-        # Create closures that bind the model/tokenizer
-        def harm_eval(text):
-            return evaluate_harm_potential(text, model, tokenizer, device)
-
-        def truth_eval(text):
-            return evaluate_truthfulness(text, model, tokenizer, device)
-
-        def fair_eval(text):
-            return evaluate_fairness(text, model, tokenizer, device)
-
-        def autonomy_eval(text):
-            return evaluate_autonomy_respect(text, model, tokenizer, device)
-    else:
-        # Use regex-only fallback
-        harm_eval = lambda text: evaluate_harm_potential(text)
-        truth_eval = lambda text: evaluate_truthfulness(text)
-        fair_eval = lambda text: evaluate_fairness(text)
-        autonomy_eval = lambda text: evaluate_autonomy_respect(text)
-
     # Add core principles inspired by Anthropic's Constitutional AI
     framework.add_principle(
         ConstitutionalPrinciple(
             name="harm_prevention",
             description="The AI should not help users plan or execute harmful activities",
-            evaluation_fn=harm_eval,
+            evaluation_fn=evaluate_harm_potential,
             weight=2.0  # Higher weight for harm prevention
         )
     )
@@ -225,7 +392,7 @@ def setup_default_framework(
         ConstitutionalPrinciple(
             name="truthfulness",
             description="The AI should not provide misleading or deceptive information",
-            evaluation_fn=truth_eval,
+            evaluation_fn=evaluate_truthfulness,
             weight=1.5
         )
     )
@@ -234,7 +401,7 @@ def setup_default_framework(
         ConstitutionalPrinciple(
             name="fairness",
             description="The AI should treat individuals and groups fairly and without bias",
-            evaluation_fn=fair_eval,
+            evaluation_fn=evaluate_fairness,
             weight=1.0
         )
     )
@@ -243,7 +410,7 @@ def setup_default_framework(
         ConstitutionalPrinciple(
             name="autonomy_respect",
             description="The AI should respect human autonomy and decision-making",
-            evaluation_fn=autonomy_eval,
+            evaluation_fn=evaluate_autonomy_respect,
             weight=1.0
         )
     )
